+++ src/components/App.routes.tsx
@@ -25,7 +25,7 @@ import Login from './routes/authentication/Login';
 import Logout from './routes/authentication/Logout';
 import SignUp from './routes/authentication/SignUp';
 import { StripeExpressComplete, StripeExpressNew } from './routes/Account/StripeExpress';
-import FirebaseAccountEmailHandler from './routes/Account/FirebaseAccountEmailHandler';
+import FirebaseEmailHandler from './routes/Account/FirebaseEmailHandler';
 import Trips from './routes/Trips';
 import TripsReceipt from './routes/Trips/TripsReceipt';
 import NotFound from './routes/NotFound';
@@ -46,7 +46,7 @@ const AppRoutes = () => (
       return (
         <AppContainer className="bee-app" {...bannerState}>
           <Switch>
-            <Route exact path="/account/action" component={FirebaseAccountEmailHandler} />
+            <Route exact path="/account/action" component={FirebaseEmailHandler} />
             <Route exact path="/account/stripe_express/new" component={StripeExpressNew} />
             <Route exact path="/account/stripe_express/complete" component={StripeExpressComplete} />
             <AuthenticatedRoute path="/account" component={Account} />
