+++ src/components/routes/Account/FirebaseEmailHandler/FirebaseEmailHandler.container.ts
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+const FirebaseEmailHandlerContainer = styled.div`
+    h2 {
+      margin-top: 18px;
+    }
+
+    section {
+      margin-top: 18px;
+    }
+
+    .input-container {
+      display: flex;
+      flex-direction: column;
+      margin: 16px 0px 28px 0px;
+      width: 100%;
+    }
+  }
+`;
+
+export default FirebaseEmailHandlerContainer;
