+++ src/components/routes/Account/FirebaseEmailHandler/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './FirebaseEmailHandler';
